+++ cmd/ftl/cmd_build.go
@@ -6,8 +6,8 @@ import (
 	"fmt"
 
 	"github.com/TBD54566975/ftl/backend/protos/xyz/block/ftl/v1/ftlv1connect"
-	"github.com/TBD54566975/ftl/common/projectconfig"
 	"github.com/TBD54566975/ftl/internal/buildengine"
+	"github.com/TBD54566975/ftl/internal/projectconfig"
 )
 
 type buildCmd struct {
