+++ cmd/ftl/cmd_init.go
@@ -9,10 +9,10 @@ import (
 	"strings"
 
 	"github.com/TBD54566975/ftl"
-	"github.com/TBD54566975/ftl/common/projectconfig"
 	"github.com/TBD54566975/ftl/internal"
 	"github.com/TBD54566975/ftl/internal/exec"
 	"github.com/TBD54566975/ftl/internal/log"
+	"github.com/TBD54566975/ftl/internal/projectconfig"
 	"github.com/TBD54566975/ftl/internal/projectinit"
 )
