+++ cmd/ftl/cmd_secret.go
@@ -13,8 +13,8 @@ import (
 	"golang.org/x/term"
 
 	ftlv1 "github.com/TBD54566975/ftl/backend/protos/xyz/block/ftl/v1"
-	"github.com/TBD54566975/ftl/common/projectconfig"
 	cf "github.com/TBD54566975/ftl/internal/configuration"
+	"github.com/TBD54566975/ftl/internal/projectconfig"
 )
 
 type secretCmd struct {
