+++ internal/projectconfig/integration_test.go
