+++ internal/projectconfig/testdata/ftl-project.toml
