+++ internal/projectconfig/testdata/go/configs-ftl-project.toml
