+++ internal/projectconfig/testdata/go/echo/echo.go
