+++ internal/projectconfig/testdata/go/echo/ftl.toml
