+++ internal/projectconfig/testdata/go/echo/go.mod
