+++ internal/projectconfig/testdata/go/findconfig/findconfig.go
