+++ internal/projectconfig/testdata/go/findconfig/ftl-project.toml
