+++ internal/projectconfig/testdata/go/findconfig/go.mod
