+++ internal/projectconfig/testdata/go/no-module-dirs-ftl-project.toml
