+++ internal/projectconfig/testdata/go/validate-configs-ftl-project.toml
