+++ internal/projectconfig/testdata/go/validateconfig/ftl-project.toml
