+++ internal/projectconfig/testdata/go/validateconfig/ftl.toml
