+++ internal/projectconfig/testdata/go/validateconfig/go.mod
