+++ internal/projectconfig/testdata/go/validateconfig/validateconfig.go
