+++ internal/projectconfig/testdata/withMinVersion/ftl-project.toml
